casting func changed

+++ superpoint/datasets/coco.py
@@ -130,11 +130,11 @@ class Coco(BaseDataset):
         if has_keypoints:
             data = data.map_parallel(pipeline.add_keypoint_map)
         data = data.map_parallel(
-            lambda d: {**d, 'image': tf.to_float(d['image']) / 255.})
+            lambda d: {**d, 'image': tf.cast(d['image'], tf.float32) / 255.})
         if config['warped_pair']['enable']:
             data = data.map_parallel(
                 lambda d: {
                     **d, 'warped': {**d['warped'],
-                                    'image': tf.to_float(d['warped']['image']) / 255.}})
+                                    'image': tf.cast(d['warped']['image'], tf.float32) / 255.}})
 
         return data
