+++ superpoint/generate_coco_patches.py
@@ -16,9 +16,9 @@ seed = None
 
 def _scale_preserving_resize(image):
     target_size = tf.convert_to_tensor(config['preprocessing']['resize'])
-    scales = tf.to_float(tf.divide(target_size, tf.shape(image)[:2]))
-    new_size = tf.to_float(tf.shape(image)[:2]) * tf.reduce_max(scales)
-    image = tf.image.resize_images(image, tf.to_int32(new_size),
+    scales = tf.cast(tf.divide(target_size, tf.shape(image)[:2]), tf.float32)
+    new_size = tf.cast(tf.shape(image)[:2]) * tf.reduce_max(scales, tf.float32)
+    image = tf.image.resize_images(image, tf.cast(new_size, tf.int32),
                                    method=tf.image.ResizeMethod.BILINEAR)
     return tf.image.resize_image_with_crop_or_pad(image, target_size[0],
                                                   target_size[1])
